Add files via upload
the board was not measured correctly. Not ready to print. Proof of concept of keeping the board on one side of the board while the launchpad is on the other side. 

No wifi board here. I need to make additional jumper holes once I know the size of the final board. Need to include spot to put esp board. We can put holes to add more stuff later.